# main.go
package main

import (
	"net/http"

	_ "docs"

	"github.com/gin-gonic/gin"
	ginSwagger "github.com/swaggo/gin-swagger"
	"github.com/swaggo/gin-swagger/swaggerFiles"
)

// swagger embed files

// album represents data about a record album.
type album struct {
	ID     string  `json:"id"`
	Title  string  `json:"title"`
	Artist string  `json:"artist"`
	Price  float64 `json:"price"`
}

// albums slice to seed record album data.
var albums = []album{
	{ID: "1", Title: "Blue Train", Artist: "John Coltrane", Price: 56.99},
	{ID: "2", Title: "Jeru", Artist: "Gerry Mulligan", Price: 17.99},
	{ID: "3", Title: "Sarah Vaughan and Clifford Brown", Artist: "Sarah Vaughan", Price: 39.99},
}

// @title User API documentation
// @version 1.0.0
// @host localhost:8080
// @BasePath /albums
func main() {
	router := gin.Default()
	router.GET("/albums/get", getAlbums)
	router.POST("/albums/post", postAlbums)
	router.GET("/albums/get/:id", getAlbumByID)

	//url := ginSwagger.URL("http://localhost:8080/swagger/swagger.json") // The url pointing to API definition
	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	router.Run(":8080")
}

// getAlbums responds with the list of all albums as JSON.
// GetAlbums ... Get all albums
// @Summary Get all albums
// @Description get all albums
// @Tags Albums
// @Success 200 {array} album
// @Failure 404 {object} object
// @Router /get/ [get]
func getAlbums(c *gin.Context) {
	c.IndentedJSON(http.StatusOK, albums)
}

// postAlbums adds an album from JSON received in the request body.
// CreateAlbum ... Create Album
// @Summary Create new album based on paramters
// @Description Create new album
// @Tags Albums
// @Accept json
// @Param newAlbum body album true "Album Data"
// @Success 201 {object} object
// @Failure 400,500 {object} object
// @Router /post/ [post]
func postAlbums(c *gin.Context) {
	var newAlbum album

	// Call BindJSON to bind the received JSON to
	// newAlbum.
	if err := c.BindJSON(&newAlbum); err != nil {
		return
	}

	// Add the new album to the slice.
	albums = append(albums, newAlbum)
	c.IndentedJSON(http.StatusCreated, newAlbum)
}

// getAlbumByID locates the album whose ID value matches the id
// parameter sent by the client, then returns that album as a response.
// GetAlbumByID ... Get the user by id
// @Summary Get one user
// @Description get album by ID
// @Tags album
// @Param id path string true "Album ID"
// @Success 200 {object} album
// @Failure 400,404 {object} object
// @Router /get/{id} [get]
func getAlbumByID(c *gin.Context) {
	id := c.Param("id")

	// Loop over the list of albums, looking for
	// an album whose ID value matches the parameter.
	for _, a := range albums {
		if a.ID == id {
			c.IndentedJSON(http.StatusOK, a)
			return
		}
	}
	c.IndentedJSON(http.StatusNotFound, gin.H{"message": "album not found"})
}